#Att 7 - Criação do modal beneficiario.

+++ FI.AtividadeEntrevista/DML/Beneficiario.cs
@@ -11,6 +11,6 @@ namespace FI.AtividadeEntrevista.DML
         public long Id { get; set; }
         public string CPF { get; set; }
         public string Nome { get; set; }
-        public int IdCliente { get; set; }
+        public long IdCliente { get; set; }
     }
 }
